+++ src/main/java/com/iexec/core/configuration/ResultRepositoryConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,27 +19,18 @@ package com.iexec.core.configuration;
 import com.iexec.resultproxy.api.ResultProxyClient;
 import com.iexec.resultproxy.api.ResultProxyClientBuilder;
 import feign.Logger;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
+import lombok.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Component;
 
-@Component
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
+@Value
+@ConstructorBinding
+@ConfigurationProperties(prefix = "result-repository")
 public class ResultRepositoryConfiguration {
-
-    @Value("${resultRepository.protocol}")
-    private String protocol;
-
-    @Value("${resultRepository.host}")
-    private String host;
-
-    @Value("${resultRepository.port}")
-    private String port;
+    String protocol;
+    String host;
+    String port;
 
     public String getResultRepositoryURL() {
         return protocol + "://" + host + ":" + port;